projectSourceCode/requirements.txt
yfinance
flask
flask-cors
psycopg2-binary